+++ src/blue_and_pink_synth_editor/ui_controls/nymphes_setup_instructions_popup.py
@@ -1,0 +1,13 @@
+from kivy.uix.boxlayout import BoxLayout
+from kivy.lang.builder import Builder
+from pathlib import Path
+
+
+# Import the kv file with the same name as this file
+parent_directory = Path(__file__).resolve().parent
+this_file_name = Path(__file__).stem
+Builder.load_file(str(parent_directory / f'{this_file_name}.kv'))
+
+
+class NymphesSetupInstructionsPopup(BoxLayout):
+    pass